move group/user hierarchy stuff into separate file
